+++ src/QueryManager.ts
@@ -58,6 +58,7 @@ import {
 
 import {
   diffSelectionSetAgainstStore,
+  removeUnusedVariablesFromQuery,
 } from './data/diffAgainstStore';
 
 import {
@@ -717,6 +718,8 @@ export class QueryManager {
         name: queryDef.name,
         fragmentMap,
       });
+
+      removeUnusedVariablesFromQuery(diffedQuery);
     }
 
     return {
@@ -725,7 +728,7 @@ export class QueryManager {
     };
   }
 
-  // Takes a request id, query id, a query document and information asscoaiated with the query
+  // Takes a request id, query id, a query document and information associated with the query
   // (e.g. variables, fragment map, etc.) and send it to the network interface. Returns
   // a promise for the result associated with that request.
   private fetchRequest({