+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Expect active development and potentially significant breaking changes in the `0
 ### vNEXT
 - Fixed an issue with named fragments in batched queries. [PR #509](https://github.com/apollostack/apollo-client/pull/509) and [Issue #501](https://github.com/apollostack/apollo-client/issues/501).
 - Fixed an issue with unused variables in queries after diffing queries against information available in the store. [PR #518](https://github.com/apollostack/apollo-client/pull/518) and [Issue #496](https://github.com/apollostack/apollo-client/issues/496).
+- Added some light type checking of variables passed to queries and mutations, avoiding accidental coercion of objects to strings. [PR #522](https://github.com/apollostack/apollo-client/pull/522) and [Issue #483](https://github.com/apollostack/apollo-client/issues/483).
 
 ### v0.4.11